Convert stokes value to casacore for PV image
Move computed stokes code to new files
