+++ CMakeLists.txt
@@ -243,8 +243,8 @@ set(SOURCE_FILES
         src/Util/App.cc
         src/Util/Casacore.cc
         src/Util/File.cc
-        src/Util/Image.cc
         src/Util/Message.cc
+		src/Util/Stokes.cc
         src/Util/String.cc
         src/Util/Token.cc
         src/ImageFitter/ImageFitter.cc)
