+++ src/Region/Region.h
@@ -21,8 +21,8 @@
 #include <casacore/lattices/Lattices/ArrayLattice.h>
 #include <casacore/tables/Tables/TableRecord.h>
 
-#include "Util/Image.h"
 #include "Util/Message.h"
+#include "Util/Stokes.h"
 
 #define DEFAULT_VERTEX_COUNT 1000
